% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DApers.R
\docType{package}
\name{DApers}
\alias{DApers}
\title{DApers: paquete para generar registros personales adecuados para analisis}
\description{
Tiene unas pocas funciones por ahora; la idea es que siga creciendo y que
sirva de inspiracion para otros paquetes que vamos a armar.
}
