% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpieza_TS_TT.R
\name{etiq_rp}
\alias{etiq_rp}
\title{Genera tabla con formato TS}
\usage{
etiq_rp(datos, tablaEtiq, nombre = "", exportar = TRUE)
}
\arguments{
\item{datos}{algo}

\item{tablaEtiq}{Tabla de Equivalencias de etiquetas TS-TT}

\item{nombre}{Nombre al nuevo data.frame}

\item{exportar}{(TRUE por defecto) si queres exportar data.frame resultante}
}
\description{
Con salida de importar_tt() se crea nuevo data.frame ajustado a formato TS
}
\details{
Es la parte 2 del proceso importar_tt() >> etiq_rp(2): crear RP20xx
}
