% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpieza_TS_TT.R
\name{hms_fch}
\alias{hms_fch}
\title{Equivalencia de fechas TS-TT}
\usage{
hms_fch(df_tt, formatoFecha = "\%d/\%m/\%Y", formatoTS = 2, exportar = TRUE)
}
\arguments{
\item{df_tt}{algo}

\item{formatoFecha}{1:version <2019, 2:version>2019 (x defecto)}

\item{formatoTS}{No usado}

\item{exportar}{(TRUE por defecto) si queres exportar data.frame resultante}
}
\description{
Cambia formato de fechas, dias, horas, etc. para hacerlo compatible con TS
}
\details{
Usada hasta 2019 con frecuencia; se deja de usar en 2021 por mejora en etiq_rp()
}
